Using new Assets client nuget

+++ src/Lykke.Service.HFT.AzureRepositories/LimitOrderStateEntity.cs
@@ -12,20 +12,20 @@ namespace Lykke.Service.HFT.AzureRepositories
         public DateTime CreatedAt { get; set; }
         public Guid Id => Guid.Parse(RowKey);
         public DateTime? LastMatchTime { get; set; }
-        public double? Price { get; set; }
+        public decimal? Price { get; set; }
         public DateTime Registered { get; set; }
-        public double RemainingVolume { get; set; }
+        public decimal RemainingVolume { get; set; }
         public OrderStatus Status { get; set; }
         public string StatusString
         {
             get => Status.ToString();
             set => Status = Enum.Parse<OrderStatus>(value);
         }
-        public double Volume { get; set; }
+        public decimal Volume { get; set; }
         public int Type { get; set; }
-        public double? LowerLimitPrice { get; set; }
-        public double? LowerPrice { get; set; }
-        public double? UpperLimitPrice { get; set; }
-        public double? UpperPrice { get; set; }
+        public decimal? LowerLimitPrice { get; set; }
+        public decimal? LowerPrice { get; set; }
+        public decimal? UpperLimitPrice { get; set; }
+        public decimal? UpperPrice { get; set; }
     }
 }
