+++ src/Lykke.Service.HFT.Core/Domain/LimitOrderState.cs
@@ -9,16 +9,16 @@ namespace Lykke.Service.HFT.Core.Domain
         public string ClientId { get; set; }
         public OrderStatus Status { get; set; }
         public string AssetPairId { get; set; }
-        public double Volume { get; set; }
-        public double? Price { get; set; }
-        public double RemainingVolume { get; set; }
+        public decimal Volume { get; set; }
+        public decimal? Price { get; set; }
+        public decimal RemainingVolume { get; set; }
         public DateTime? LastMatchTime { get; set; }
         public DateTime CreatedAt { get; set; }
         public DateTime Registered { get; set; }
         public int Type { get; set; } = 0;
-        public double? LowerLimitPrice { get; set; }
-        public double? LowerPrice { get; set; }
-        public double? UpperLimitPrice { get; set; }
-        public double? UpperPrice { get; set; }
+        public decimal? LowerLimitPrice { get; set; }
+        public decimal? LowerPrice { get; set; }
+        public decimal? UpperLimitPrice { get; set; }
+        public decimal? UpperPrice { get; set; }
     }
 }
